feat: 날짜 정렬 로직을 컴포넌트 내부에서 구현
