+++ src/pages/DiaryList/DiaryList.jsx
@@ -5,21 +5,15 @@ import { diaryListAtom } from '../../recoil/atoms';
 import { useNavigate } from 'react-router-dom';
 import BtnBack from '../../components/common/buttons/Back/BtnBack';
 import DiaryListComponent from '../../components/common/DiaryListComponent/DiaryListComponent';
-import useParsingDate from '../../hooks/diary/useParsingDate';
 
 export default function DiaryList(){
   const [diaryList, setDiaryList] = useRecoilState(diaryListAtom);
   const navigate = useNavigate();
-  const parsingDate = useParsingDate();
 
   const handleBackButton = () => {
     setDiaryList({});
     navigate('/searchbyemotion');
   };
-
-  useEffect(() => {
-    parsingDate();
-  }, []);
 
   console.log(diaryList)
   return(
