+++ src/pages/SearchByEmotion/SearchByEmotionPopUp/SearchByEmotionPopUp.jsx
@@ -15,6 +15,23 @@ export default function SearchByEmotionPopUp(){
   const mutation = usePostFeelingList();
   const navigate = useNavigate();
   
+  const handleDate = (list) => {
+    list.map((item) => {
+      const date = new Date(item.createdDate);
+      const year = date.getFullYear();
+      const month = date.getMonth() + 1;
+      const day = date.getDate();
+      const newDate = `${year}년 ${month}월 ${day}일`;
+      
+      item.createdDate = newDate;
+
+    });
+    
+    return list;
+  };
+
+
+
   const handleClick = (btnEnum) => {
     const body = {
       feeling : btnEnum,
@@ -22,7 +39,8 @@ export default function SearchByEmotionPopUp(){
 
     mutation.mutate(body, {
       onSuccess: (response) => {
-        setDiaryList(response.data.feelingList);  
+
+        setDiaryList(handleDate(response.data.feelingList));  
         navigate('/searchbyemotion/diarylist');
       },
       onError: (error) => {
